+++ ckanext/dcat/utils.py
@@ -434,6 +434,10 @@ def read_catalog_page(_format):
     data_dict = {
         'page': toolkit.request.params.get('page'),
         'modified_since': toolkit.request.params.get('modified_since'),
+        # param with organization defined
+        'organization': toolkit.request.params.get('organization'),
+        # param to overwrite/bulk patch original metadata values
+        'overwrite_values': toolkit.request.params.get('overwrite'),
         'q': toolkit.request.params.get('q'),
         'fq': toolkit.request.params.get('fq'),
         'format': _format,